+++ packages/extension/src/ui/features/accounts/AccountListScreen.tsx
@@ -13,6 +13,7 @@ import styled from "styled-components"
 import { routes, useReturnTo } from "../../routes"
 import { P } from "../../theme/Typography"
 import { LoadingScreen } from "../actions/LoadingScreen"
+import { MultisigListScreenItem } from "../multisig/MultisigListScreenItem"
 import { useCurrentNetwork } from "../networks/useNetworks"
 import { useBackupRequired } from "../recovery/backupDownload.state"
 import { recover } from "../recovery/recovery.service"
@@ -26,7 +27,6 @@ import {
 import { DeprecatedAccountsWarning } from "./DeprecatedAccountsWarning"
 import { GroupedAccountList } from "./GroupedAccountList"
 import { HiddenAccountsBar } from "./HiddenAccountsBar"
-import { MultisigListScreenItem } from "./multisig/MultisigListScreenItem"
 import { usePartitionDeprecatedAccounts } from "./upgrade.service"
 import { useAddAccount } from "./useAddAccount"
