+++ packages/extension/src/ui/features/multisig/CreateMultisigScreen/MultisigSecondStep.tsx
@@ -3,10 +3,10 @@ import { icons } from "@argent/ui"
 import { Box, Button, Center, Flex } from "@chakra-ui/react"
 import { Controller, useFormContext } from "react-hook-form"
 
-import { isEmptyValue } from "../../../../../shared/utils/object"
-import { useAppState } from "../../../../app.state"
-import { useSelectedAccount } from "../../accounts.state"
-import { useNextPublicKey, useNextSignerKey } from "../../usePublicKey"
+import { isEmptyValue } from "../../../../shared/utils/object"
+import { useAppState } from "../../../app.state"
+import { useSelectedAccount } from "../../accounts/accounts.state"
+import { useNextPublicKey, useNextSignerKey } from "../../accounts/usePublicKey"
 import { ScreenLayout } from "./ScreenLayout"
 import { useCreateMultisig } from "./useCreateMultisig"
 import { FieldValues } from "./useCreateMultisigForm"
