+++ packages/extension/src/ui/features/multisig/CreateMultisigScreen/ScreenLayout.tsx
@@ -3,10 +3,10 @@ import { Box, Button } from "@chakra-ui/react"
 import { isNumber } from "lodash-es"
 import { FC, PropsWithChildren, ReactNode } from "react"
 
-import { ContentWrapper } from "../../../../components/FullScreenPage"
-import { ArrowBackIcon } from "../../../../components/Icons/MuiIcons"
-import { StepIndicator } from "../../../../components/StepIndicator"
-import LogoSvg from "../../../lock/logo.svg"
+import { ContentWrapper } from "../../../components/FullScreenPage"
+import { ArrowBackIcon } from "../../../components/Icons/MuiIcons"
+import { StepIndicator } from "../../../components/StepIndicator"
+import LogoSvg from "../../lock/logo.svg"
 
 export interface CreateMultisigScreen extends PropsWithChildren {
   back?: boolean
