+++ packages/extension/src/ui/features/multisig/MultisigDeleteModal.tsx
