+++ packages/extension/src/ui/features/multisig/hooks/useMultisigStatus.ts
@@ -1,0 +1,18 @@
+import { BaseWalletAccount } from "../../../../shared/wallet.model"
+import { isZeroMultisigAccount, useMultisigAccount } from "../multisig.state"
+
+export type MultisigStatus = "pending" | "ready" | "unknown"
+
+export function useMultisigStatus(account: BaseWalletAccount): MultisigStatus {
+  const multisigAccount = useMultisigAccount(account)
+
+  if (!multisigAccount) {
+    return "unknown"
+  }
+
+  if (isZeroMultisigAccount(multisigAccount)) {
+    return "pending"
+  }
+
+  return "ready"
+}
